clippy

+++ examples/minimal.rs
@@ -80,8 +80,8 @@ fn intersection(query: Query<&Intersection<MyRaycastSet>>) {
 fn rotator(time: Res<Time>, mut query: Query<&mut Transform, With<RaycastSource<MyRaycastSet>>>) {
     for mut transform in &mut query {
         *transform = Transform::from_rotation(
-            Quat::from_rotation_x(time.elapsed_seconds().sin() as f32 * 0.2)
-                * Quat::from_rotation_y((time.elapsed_seconds() * 1.5).sin() as f32 * 0.1),
+            Quat::from_rotation_x(time.elapsed_seconds().sin() * 0.2)
+                * Quat::from_rotation_y((time.elapsed_seconds() * 1.5).sin() * 0.1),
         );
     }
 }
